pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-manager"
version = "0.1.0"
description = "Cross-platform photo organizer and viewer"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0",
    "imagehash>=4.3",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
